+++ packages/happn-cluster/RELEASES.md
@@ -340,3 +340,4 @@ non-release 2017-12-14
 12.3.0 2023-01-03
 -----------------
   - feat: Summon productionization
+  - fix: SQLite nested ands and ors